fix(CANDriver): type correctness and stack safety improvements
- Change dlc parameter and txDlc member from int/int to int8_t
  (sentinel -1 and values 0..8 all fit; saves 3 bytes per instance)
- Add static to cnfMapper in MCP2515::begin() — without it the 264-byte
  table was allocated on the AVR stack each call (potential overflow)
- Cast txId to uint32_t before bit-shifts in MCP2515 and ESP32SJA1000
  endPacket() to avoid shifting a signed integer (technically UB)
- Remove redundant casts in endPacket() where txLength is already uint8_t

+++ lib/CANDriver/src/CANController.cpp
@@ -38,7 +38,7 @@ uint8_t CANController::begin(uint32_t /*baudRate*/) {
 
 void CANController::end() {}
 
-uint8_t CANController::beginPacket(uint16_t id, int dlc, bool rtr) {
+uint8_t CANController::beginPacket(uint16_t id, int8_t dlc, bool rtr) {
   if(id > 0x7FFU) { return 0U; }
   if(dlc > 8) { return 0U; }
 
@@ -54,7 +54,7 @@ uint8_t CANController::beginPacket(uint16_t id, int dlc, bool rtr) {
   return 1U;
 }
 
-uint8_t CANController::beginExtendedPacket(uint32_t id, int dlc, bool rtr) {
+uint8_t CANController::beginExtendedPacket(uint32_t id, int8_t dlc, bool rtr) {
   if(id > 0x1FFFFFFFU) { return 0U; }
   if(dlc > 8) { return 0U; }
 
@@ -75,7 +75,7 @@ uint8_t CANController::endPacket() {
   packetBegun = false;
 
   if(txDlc >= 0) {
-    txLength = static_cast<uint8_t>(txDlc);
+    txLength = static_cast<uint8_t>(txDlc); // txDlc is 0..8 here
   }
 
   return 1U;
